pyproject.toml
[project]
name = "opkey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.4.0",
    "mcp[cli]>=1.9.1",
    "python-dotenv>=1.1.0",
]
